requirements.txt
backtesting==0.6.5
bokeh==3.9.0
boto3==1.42.72
botocore==1.42.72
contourpy==1.3.3
Jinja2==3.1.6
jmespath==1.1.0
MarkupSafe==3.0.3
narwhals==2.18.0
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pillow==12.1.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
s3transfer==0.16.0
six==1.17.0
tornado==6.5.5
urllib3==2.6.3
xyzservices==2025.11.0
